Remove comments from fast-export AST
The fast-export-rust AST started intended as a concrete syntax tree,
preserving all trivia, so it could be a 1-to-1 representation and
represent all inputs accepted by fast-import. I now think that is not
important. None of that affects the repo itself.

There are many equally valid representations, such as optional LFs,
positive-signed numbers, leading zeros, and uppercase and lowercase file
size units, that just don't matter and would needlessly complicate the
representation. Rather, I will use a canonical pretty printing, that
matches fast-export. If it ever matters for textual comparisons, the
streams could be canonicalized with a simple parse-and-pretty-print
loop.

Comments are more impactful than the other trivia. The now-removed setup
wraps fields in `Commented`. This elegantly keeps the comments
associated with the following lines, but it's expensive, as it adds 24
bytes to every field. I assume comments are rare, likely written by hand
in small test cases, though perhaps some exporters use them as ad hoc
commands or metadata. A possible solution could be to group all comments
for a command into one `Vec<Option<Comment>>` with a length matching the
number of lines in the command, but this leaves room for the comment to
become out of sync with their associated lines. Since comments have no
impact on the repo with fast-import, I am leaving them out.