Skip data streams without processing
If the user does not want to process a data stream, they can call
`DataReader::skip_rest`, so that the next call to `Parser::next` will
skip it without coping it to some auxiliary buffer.

This also moves the body of `DataReader::read_next` into `Parser`, but
require the caller to prove exclusivity.