Handle optional trailing LF
This handles an optional LF, which can follow several commands. If it
was handled afterwards, it could appear after a data stream. Then, in
the case it isn't followed by an empty line, it would need to retain the
line or byte until the next call. It's easier to check at the start, and
skip the newly read line if blank.