Support any byte type in Command
This changes commands to be able to store any type for their bytes
fields, so that the user doesn't need to make their own command types.

The parser still returns `&[u8]` instead of a generic `FromBytes`,
because allowing any return type has significant downsides. The lifetime
of `FromBytes` is tied to the call of `Parser::next`, so it cannot be
specified in `Parser` itself (then `next` could only be called once).
This leads to it often needing to be specified explicitly, when it
cannot be inferred, and when `Parser::next` is called with multiple
types, the large parsing methods are duplicated from monomorphism, all
just for some small conversion functions at the returns. Really, the
parser just deals with byte slices, not your type.

Keeping the `FromBytes` conversion out of the parser actually gives the
caller more flexibility too. They can selectively choose when/if to do
the conversion, based on the type of command or any other information.
This keeps the parser consistent with pull parsing instead of push
parsing.