Parse commit metadata
Commit messages are a tricky case. They appear before the `from` and
`merge` commands, instead of at the end of the command, yet they can be
arbitrary length. fast-import buffers them into memory, regardless of
the length.

Since I think it's unlikely most cases of rewriting the commit message
will depend on `from`, `merge`, and the change list, I am keeping it as
a stream for now.