License under GPL 2.0 with a linking exception
Previously, git-transform-repo was licensed under either MIT or GPL 2.0
with a linking exception. That was chosen to be compatible with
git-filter-repo, which is distributed under MIT, GPL 2.0 (though not
explicitly with a linking exception), or any license used now or in the
future by git or libgit2.

However, fast-export-rust directly references git fast-export in its
parser implementation, not with a clean-room design, so it may be
considered a derivative work of git, as GPL 2.0 defines a “translat[ion]
into another language” as a derivative work. Using GPL 2.0 with the
linking exception of libgit2 thus seems to have the best compatibility
between the three and has the benefit of encouraging improvements to be
upstreamed, while allowing unrestricted library usage.