Switch from a single buffer to windows of line buffers
`Parser::command_buf` was a single buffer which accumulates all the
lines for a single command. Since reading another line would invalidate
slices to it for previous lines, this required all parsing to be
performed in terms of spans. This also lead to a bifurcation of command
types for `&[u8]` and `Span`, either generic over the bytes type or with
separate types.

To allow reading of blob data streams to be controlled by the caller, a
single `DataReader` instance has exclusive access to the `Parser`, and
this worked because the command buffer and data state are disjoint.

Parsing commit changes is more complicated under that scheme and
necessitated this change. The planned design is to expose an iterator
of changes, including blob data streams and queries like `cat-blob` and
`ls`, as well as collect the changes (without blob data or queries) into
a `Vec` so they can be more easily manipulated. This means all change
directives need to be retained until all changes have been read. If it
were done with a single `change_buf`, all the parsing utilities would
need to be duplicated between the parser and change iterator, but under
a shared reference. It would be messy.

Instead, this change switches to separate buffers per line. Crucially,
once a buffer is pushed, its contents and allocation cannot be modified,
so any slices to it are stable. All of the lines for the current command
are retained, as well as 20 lines from before for context (as opposed to
git fast-export, which has the latest 100 lines). Thus, slices can be
manipulated and retained in the usual way and `&mut self` methods force
the caller to release them before they would be invalidated.