Rename sub-command to directive
A line is a directive, #-comment, or a line of delimited data. A
directive is a field in a command, which spans a single line. A command
is a list of associated directives, that form one logical structure.

git fast-import and fast-export docs use the term “directive” a few
times, but usually write “command” for both terms.